requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.2
httpx==0.27.0
langchain==0.1.13
langchain-openai==0.1.1
langgraph==0.0.30
python-dotenv==1.0.1
pydantic==2.6.4
matplotlib==3.8.4
plotly==5.15.0